Fix: if GC not already triggered, session valid
The GC will be triggered randomly based on gc.divisor-values etc... if
you have low traffic on your page, the gc won't be executed in a
continous manner, so that the session will be valid, even if it's
lifetime is more than "maxlifetime".

The fix checks condition even in read-out of sessions.