Drops isDestroyed flag, drops check for empty data
After destroying a session you might want to restart it again, the current
implementation doesn't allow it unless you instantiate the adapter again.
Removing the isDestroyed flag seems to fix the issue. Also, removing the
empty data check on writing allows the session data to be cleared and
fixes the bug where data was getting stuck in the session even after
deleting it.