Fixes Database::destroy
The call to session_regenerate_id has no place in the destroy method, it can actually mess things up.
Calling the destroy method on the adapter doesn't seem to have the same
result as when calling session_destroy. Since we usually use the adapter
directly, a call to session_destroy is needed.