Fix deleteNode BUG
When I use deleteNode, I encountered this error  
You should not use Sl\Models\Areas:beforeUpdate when Phalcon\Mvc\Model\Behavior\NestedSet attached. Use the methods of behavior.
I found that there is a problem in line 581, I am not sure if it is BUG, but I removed it, it worked.